[Feat]
- main 안에 들어가는 코드 App으로 대거 분리
- 인증된 사용자도 새로고침 시 로그인 페이지로 넘어가는 버그 수정
